rename to go-music-theory
