+++ chord/chord_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/yaml.v2"
 
-	"github.com/go-music/music/theory/key"
-	"github.com/go-music/music/theory/note"
+	"github.com/go-music-theory/music-theory/key"
+	"github.com/go-music-theory/music-theory/note"
 )
 
 func TestChordExpectations(t *testing.T) {
