+++ chord/function_test.go
