+++ chord/technique.go
