+++ chord/technique_test.go
