+++ key/key.go
@@ -2,7 +2,7 @@
 package key
 
 import (
-	"github.com/go-music/music/theory/note"
+	"github.com/go-music-theory/music-theory/note"
 	"regexp"
 )
