+++ note/README.md
@@ -1,0 +1,9 @@
+# Note
+
+[![GoDoc](https://godoc.org/github.com/go-music-theory/music-theory/note?status.svg)](https://godoc.org/github.com/go-music-theory/music-theory/note) [![Coverage](https://img.shields.io/badge/coverage-100%-brightgreen.svg?style=flat)](https://gocover.io/github.com/go-music-theory/music-theory/note)
+
+#### A model of a musical note.
+
+A Note is used to represent the relative duration and pitch of a sound.
+
+Author: [Charney Kaye](http://w.charney.io)
