+++ note/class.go
