+++ note/class_test.go
