+++ note/note.go
