+++ note/note_test.go
