+++ note/root.go
