--- theory/chord/README.md
@@ -1,9 +1,0 @@
-# Chord
-
-[![GoDoc](https://godoc.org/github.com/go-music/music/theory/chord?status.svg)](https://godoc.org/github.com/go-music/music/theory/chord) [![Coverage](https://img.shields.io/badge/coverage-100%-brightgreen.svg?style=flat)](https://gocover.io/github.com/go-music/music/theory/chord)
-
-#### A model of a musical chord.
-
-A chord, in music, is any harmonic set of three or more notes that is heard as if sounding simultaneously.
-
-Author: [Charney Kaye](http://w.charney.io)
