--- theory/key/README.md
@@ -1,9 +1,0 @@
-# Key
-
-[![GoDoc](https://godoc.org/github.com/go-music/music/theory/key?status.svg)](https://godoc.org/github.com/go-music/music/theory/key) [![Coverage](https://img.shields.io/badge/coverage-100%-brightgreen.svg?style=flat)](https://gocover.io/github.com/go-music/music/theory/key)
-
-#### A model of a musical key signature.
-
-The key of a piece is a group of pitches, or scale upon which a music composition is created in classical, Western art, and Western pop music.
-
-Author: [Charney Kaye](http://w.charney.io)
